Changed device database to reference token instead of user_id as we only want to receive notifications while logged in. Also added function calls to login/logout to register devices #implement

+++ server/src/main/java/seng302/Logic/Database/Notifications.java
@@ -14,17 +14,18 @@ public class Notifications {
     /**
      * Register a mobile device against a user id
      * @param device_id The Visual Studio App Center device UUID which identifies the device
-     * @param user_id The user id which identifies the user
+     * @param token The login token of the user
      * @throws SQLException When something goes wrong
      */
-    public void register(String device_id, String user_id) throws SQLException {
+    public void register(String device_id, String token) throws SQLException {
+        if(device_id == null) return;
         try(Connection connection = DatabaseConfiguration.getInstance().getConnection()) {
             //
-            String query = "INSERT INTO PUSH_DEVICE (device_id, user_id) VALUES(?, ?) ON DUPLICATE KEY UPDATE user_id=?";
+            String query = "INSERT INTO PUSH_DEVICE (device_id, user_token) VALUES(?, ?) ON DUPLICATE KEY UPDATE user_id=?";
             PreparedStatement statement = connection.prepareStatement(query);
-
             statement.setString(1, device_id);
-            statement.setString(2, user_id);
+            statement.setString(2, token);
+            statement.setString(3, device_id);
             statement.execute();
         }
     }
@@ -35,6 +36,7 @@ public class Notifications {
      * @throws SQLException When something goes wrong
      */
     public void unregister(String device_id) throws SQLException {
+        if(device_id == null) return;
         try(Connection connection = DatabaseConfiguration.getInstance().getConnection()) {
             //
             String query = "DELETE FROM PUSH_DEVICE WHERE device_id = ?";
@@ -48,17 +50,17 @@ public class Notifications {
 
     /**
      * Get all the devices on which a user is registered to receive push notifications
-     * @param user_id The id of the user
+     * @param token The login token of the user
      * @return A list of the device UUIDs on which the user has logged in
      * @throws SQLException When something goes wrong
      */
-    public List<String> getDevices(String user_id) throws SQLException {
+    public List<String> getDevices(String token) throws SQLException {
         try(Connection connection = DatabaseConfiguration.getInstance().getConnection()) {
             //
-            String query = "SELECT device_id FROM PUSH_DEVICE WHERE user_id = ?";
+            String query = "SELECT device_id FROM PUSH_DEVICE WHERE user_token = ?";
             PreparedStatement statement = connection.prepareStatement(query);
 
-            statement.setString(1, user_id);
+            statement.setString(1, token);
             ResultSet resultSet = statement.executeQuery();
             List<String> devices = new ArrayList<>();
             while(resultSet.next()) {
